backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Config & validation
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
pyyaml==6.0.2

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0          # PostgreSQL async driver
aiosqlite==0.20.0        # SQLite for local dev

# HTTP client (for platform APIs)
httpx==0.28.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19

# WebSocket
websockets==14.1

# Claude AI
anthropic==0.40.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Dev / testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1            # also used for test client
